FossilOrigin-Name: 0ffacb525a925ba8da3152c9041fdf98e67b14f10faf30dc685dea5a8ab4e9c6

+++ src/th_lang.c
@@ -958,7 +958,7 @@ static int string_length_command(
 static int string_match_command(
   Th_Interp *interp, void *ctx, int argc, const char **argv, int *argl
 ){
-  extern char *fossil_strndup(const char*,int);
+  extern char *fossil_strndup(const char*,ssize_t);
   extern void fossil_free(void*);
   char *zPat, *zStr;
   int rc;
